traffic_api/requirements.txt
asgiref==3.6.0
Django==3.2.18
djangorestframework==3.14.0
djangorestframework-gis==1.0
numpy==1.21.6
pandas==1.3.5
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
